Update ex4-4.md
- Adds code markers to `bind`, `return`, `Gen`, `Maybe`, `[]` and `MCPrelude`.
- Changes two (2) to three (3) instances. See discussion below.

I got hung up on the `two` here. When completing/finishing this challenge one should have three (3) instances (`Gen`, `Maybe`, and `[]`). Even though only though the next paragraph talks about `Gen`.

So I'd propose to either change it to three, as done in this PR. Or change the wording to say

> When you are finished with this paragraph ...

I did not equate paragraphs to challenges. Challenges for me were the whole exercise.