Removes 'Gen' from signature of genTwo
According to the description, genTwo should take the result of the first generator to return a second generator. The output function requires one seed to pass to (Gen a) but we do not require a second seed for (Gen b) because we take it from the results of (Gen a). Therefore, the projection function should not have a return type that expects a second seed.